+++ core/src/com/rock/golf/Pathfinding/RouteNode.java
@@ -3,19 +3,19 @@ package com.rock.golf.Pathfinding;
 public class RouteNode<Node> extends Graph {
     final Node current_node;
     private Node previous_node;
-    private static int routeScore;
+    private static double routeScore;
 
-    RouteNode(Node current, Node previous, int routeScore) {
+    RouteNode(Node current, Node previous, double routeScore) {
         current_node = current;
         previous_node = previous;
         this.routeScore = routeScore;
     }
 
-    public void setRouteScore(int score) {
+    public void setRouteScore(double score) {
         routeScore = score;
     }
 
-    public int getRouteSxore() {
+    public double getRouteSxore() {
         return routeScore;
     }
 }
